refactor: extract PrismaExtension
